Move join code to contrib module join-library. Adding pom.xml and README.md.
