+++ contrib/join-library/src/test/java/com/google/cloud/dataflow/contrib/joinlibrary/OuterLeftJoinTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.dataflow.sdk.transforms.join;
+package com.google.cloud.dataflow.contrib.joinlibrary;
 
 import com.google.cloud.dataflow.sdk.Pipeline;
 import com.google.cloud.dataflow.sdk.testing.DataflowAssert;
@@ -31,9 +31,9 @@ import java.util.List;
 
 
 /**
- * This test Outer Right Join functionality.
+ * This test Outer Left Join functionality.
  */
-public class OuterRightJoinTest {
+public class OuterLeftJoinTest {
 
   Pipeline p;
   List<KV<String, Long>> leftListOfKv;
@@ -60,8 +60,8 @@ public class OuterRightJoinTest {
     listRightOfKv.add(KV.of("Key2", "bar"));
     PCollection<KV<String, String>> rightCollection = p.apply(Create.of(listRightOfKv));
 
-    PCollection<KV<String, KV<Long, String>>> output = Join.rightOuterJoin(
-      leftCollection, rightCollection, -1L);
+    PCollection<KV<String, KV<Long, String>>> output = Join.leftOuterJoin(
+      leftCollection, rightCollection, "");
 
     expectedResult.add(KV.of("Key1", KV.of(5L, "foo")));
     expectedResult.add(KV.of("Key2", KV.of(4L, "bar")));
@@ -79,8 +79,8 @@ public class OuterRightJoinTest {
     listRightOfKv.add(KV.of("Key2", "gazonk"));
     PCollection<KV<String, String>> rightCollection = p.apply(Create.of(listRightOfKv));
 
-    PCollection<KV<String, KV<Long, String>>> output = Join.rightOuterJoin(
-      leftCollection, rightCollection, -1L);
+    PCollection<KV<String, KV<Long, String>>> output = Join.leftOuterJoin(
+      leftCollection, rightCollection, "");
 
     expectedResult.add(KV.of("Key2", KV.of(4L, "bar")));
     expectedResult.add(KV.of("Key2", KV.of(4L, "gazonk")));
@@ -98,8 +98,8 @@ public class OuterRightJoinTest {
     listRightOfKv.add(KV.of("Key2", "bar"));
     PCollection<KV<String, String>> rightCollection = p.apply(Create.of(listRightOfKv));
 
-    PCollection<KV<String, KV<Long, String>>> output = Join.rightOuterJoin(
-      leftCollection, rightCollection, -1L);
+    PCollection<KV<String, KV<Long, String>>> output = Join.leftOuterJoin(
+      leftCollection, rightCollection, "");
 
     expectedResult.add(KV.of("Key2", KV.of(4L, "bar")));
     expectedResult.add(KV.of("Key2", KV.of(6L, "bar")));
@@ -109,33 +109,33 @@ public class OuterRightJoinTest {
   }
 
   @Test
-  public void testJoinNoneToOneMapping() {
+  public void testJoinOneToNoneMapping() {
     leftListOfKv.add(KV.of("Key2", 4L));
     PCollection<KV<String, Long>> leftCollection = p.apply(Create.of(leftListOfKv));
 
     listRightOfKv.add(KV.of("Key3", "bar"));
     PCollection<KV<String, String>> rightCollection = p.apply(Create.of(listRightOfKv));
 
-    PCollection<KV<String, KV<Long, String>>> output = Join.rightOuterJoin(
-      leftCollection, rightCollection, -1L);
+    PCollection<KV<String, KV<Long, String>>> output = Join.leftOuterJoin(
+      leftCollection, rightCollection, "");
 
-    expectedResult.add(KV.of("Key3", KV.of(-1L, "bar")));
+    expectedResult.add(KV.of("Key2", KV.of(4L, "")));
     DataflowAssert.that(output).containsInAnyOrder(expectedResult);
     p.run();
   }
 
   @Test(expected = NullPointerException.class)
   public void testJoinLeftCollectionNull() {
-    Join.rightOuterJoin(null, p.apply(Create.of(listRightOfKv)), "");
+    Join.leftOuterJoin(null, p.apply(Create.of(listRightOfKv)), "");
   }
 
   @Test(expected = NullPointerException.class)
   public void testJoinRightCollectionNull() {
-    Join.rightOuterJoin(p.apply(Create.of(leftListOfKv)), null, -1L);
+    Join.leftOuterJoin(p.apply(Create.of(leftListOfKv)), null, "");
   }
 
   @Test(expected = NullPointerException.class)
   public void testJoinNullValueIsNull() {
-    Join.rightOuterJoin(p.apply(Create.of(leftListOfKv)), p.apply(Create.of(listRightOfKv)), null);
+    Join.leftOuterJoin(p.apply(Create.of(leftListOfKv)), p.apply(Create.of(listRightOfKv)), null);
   }
 }
